# Faster summarise_draws