--- a/R/rvar-slice.R
+++ b/R/rvar-slice.R
@@ -138,10 +138,9 @@
       if (all(names(new_dimnames) == "")) names(new_dimnames) <- NULL
       dimnames(x) <- new_dimnames
     }
-    x
-  } else {
-    x
   }
+
+  x
 }
 
 #' @export