This can close https://github.com/stan-dev/posterior/issues/98.

`summarise_draws` now internally writes output to a preallocated matrix, which makes the "chunking" option unnecessary except for parallelization, which will be dealt with separately.  

Additionally, we now check whether
```
  if ("variable" %in% the_names) {
    stop2("Name 'variable' is reserved in 'summarise_draws'.")
  }
```
prior to the potentially lengthy computation.

Benchmarks are:

```
remotes::install_github("jsocolar/posterior")
library(posterior)
library(microbenchmark)

set.seed(1)
nc <- 4

n <- 100
test_array <- array(data = rnorm(1000*nc*n), dim = c(1000,nc,n))
y <- posterior::as_draws_array(test_array)

b100 <- microbenchmark(m_s <- posterior::summarise_draws(y), times = 10)
# .786 sec under current posterior
# .79 sec under new version

n <- 1000
test_array <- array(data = rnorm(1000*nc*n), dim = c(1000,nc,n))
y <- posterior::as_draws_array(test_array)

b1000 <- microbenchmark(m_s <- posterior::summarise_draws(y), times = 2)
# 7.5 sec under current posterior
# 7.5 sec under new version

n <- 10000
test_array <- array(data = rnorm(1000*nc*n), dim = c(1000,nc,n))
y <- posterior::as_draws_array(test_array)

b10000 <- microbenchmark(m_s <- posterior::summarise_draws(y), times = 1)
# 84 under current posterior
# 73.5 sec under new version

n <- 50000
test_array <- array(data = rnorm(1000*nc*n), dim = c(1000,nc,n))
y <- posterior::as_draws_array(test_array)

b50000 <- microbenchmark(m_s <- posterior::summarise_draws(y), times = 1)
# 701 sec under current posterior
# 385 sec under new version
```